# src/IO.Swagger/Models/Mailing.cs
/*
 * MailChampApi
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace IO.Swagger.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class Mailing : IEquatable<Mailing>
    { 
        /// <summary>
        /// Gets or Sets MailingId
        /// </summary>

        [DataMember(Name="mailingId")]
        public int? MailingId { get; set; }

        /// <summary>
        /// Gets or Sets MailingName
        /// </summary>

        [DataMember(Name="mailingName")]
        public string MailingName { get; set; }

        /// <summary>
        /// Gets or Sets DateCreated
        /// </summary>

        [DataMember(Name="dateCreated")]
        public DateTime? DateCreated { get; set; }

        /// <summary>
        /// Gets or Sets DateModified
        /// </summary>

        [DataMember(Name="dateModified")]
        public DateTime? DateModified { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class Mailing {\n");
            sb.Append("  MailingId: ").Append(MailingId).Append("\n");
            sb.Append("  MailingName: ").Append(MailingName).Append("\n");
            sb.Append("  DateCreated: ").Append(DateCreated).Append("\n");
            sb.Append("  DateModified: ").Append(DateModified).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((Mailing)obj);
        }

        /// <summary>
        /// Returns true if Mailing instances are equal
        /// </summary>
        /// <param name="other">Instance of Mailing to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(Mailing other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    MailingId == other.MailingId ||
                    MailingId != null &&
                    MailingId.Equals(other.MailingId)
                ) && 
                (
                    MailingName == other.MailingName ||
                    MailingName != null &&
                    MailingName.Equals(other.MailingName)
                ) && 
                (
                    DateCreated == other.DateCreated ||
                    DateCreated != null &&
                    DateCreated.Equals(other.DateCreated)
                ) && 
                (
                    DateModified == other.DateModified ||
                    DateModified != null &&
                    DateModified.Equals(other.DateModified)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (MailingId != null)
                    hashCode = hashCode * 59 + MailingId.GetHashCode();
                    if (MailingName != null)
                    hashCode = hashCode * 59 + MailingName.GetHashCode();
                    if (DateCreated != null)
                    hashCode = hashCode * 59 + DateCreated.GetHashCode();
                    if (DateModified != null)
                    hashCode = hashCode * 59 + DateModified.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(Mailing left, Mailing right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(Mailing left, Mailing right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}