# src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "MailChampApi",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
    "/Mailings/{mailingId}" : {
      "get" : {
        "tags" : [ "Mailings" ],
        "summary" : "Gets the requested mailing.",
        "parameters" : [ {
          "name" : "mailingId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Mailing"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Mailing"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Mailing"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Mailings" : {
      "post" : {
        "tags" : [ "Mailings" ],
        "summary" : "Creates a new mailing.",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MailingInput"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MailingInput"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MailingInput"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Mailing"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Mailing"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Mailing"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Mailings/{mailingId}/name" : {
      "put" : {
        "tags" : [ "Mailings" ],
        "summary" : "Updates the name of the selected mailing.",
        "parameters" : [ {
          "name" : "mailingId",
          "in" : "path",
          "description" : "The mailing ID of the mailing to update.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "description" : "The mailing ID of the mailing to update.",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "description" : "The new mailing name.",
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "string",
                "description" : "The new mailing name.",
                "nullable" : true
              }
            },
            "text/json" : {
              "schema" : {
                "type" : "string",
                "description" : "The new mailing name.",
                "nullable" : true
              }
            },
            "application/*+json" : {
              "schema" : {
                "type" : "string",
                "description" : "The new mailing name.",
                "nullable" : true
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Mailing"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Mailing"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Mailing"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Mailings/{mailingId}/send" : {
      "post" : {
        "tags" : [ "Mailings" ],
        "summary" : "Queues the selected mailing to be sent for each of the provided records.\r\nIt is recommended to include no more than 1024 records per request.",
        "parameters" : [ {
          "name" : "mailingId",
          "in" : "path",
          "description" : "The mailing ID of the mailing to send.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "description" : "The mailing ID of the mailing to send.",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "description" : "A collection of records, each consisting of a collection \r\nof key-value pairs to be used in the generation of each email.",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SendInput"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SendInput"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SendInput"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Mailing" : {
        "type" : "object",
        "properties" : {
          "mailingId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "mailingName" : {
            "type" : "string",
            "nullable" : true
          },
          "dateCreated" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dateModified" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "additionalProperties" : false,
        "example" : {
          "dateCreated" : "2000-01-23T04:56:07.000+00:00",
          "dateModified" : "2000-01-23T04:56:07.000+00:00",
          "mailingId" : 0,
          "mailingName" : "mailingName"
        }
      },
      "ProblemDetails" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "nullable" : true
          },
          "title" : {
            "type" : "string",
            "nullable" : true
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          },
          "detail" : {
            "type" : "string",
            "nullable" : true
          },
          "instance" : {
            "type" : "string",
            "nullable" : true
          }
        },
        "additionalProperties" : { }
      },
      "MailingInput" : {
        "type" : "object",
        "properties" : {
          "mailingName" : {
            "type" : "string",
            "nullable" : true
          }
        },
        "additionalProperties" : false
      },
      "SendInput" : {
        "type" : "object",
        "properties" : {
          "records" : {
            "type" : "array",
            "nullable" : true,
            "items" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "string"
              }
            }
          }
        },
        "additionalProperties" : false
      },
...
    }
  }
}